+++ gobang_server/ServerSocket.h
@@ -27,9 +27,9 @@ public:
 
 private:
 	//表示服务器是否运行的变量
-	int running;
+	uint16_t running;
 	//下一个客户端连进来时的index
-	int next_ind;
+	uint16_t next_ind;
 	//保存服务器的socket变量
 	TCPsocket server_socket;
 	//客户端数组， clients和sockets一一对应
@@ -42,17 +42,17 @@ private:
 	//初始化
 	void init();
 	//当有socket流入时调用
-	int acceptSocket(int index);
+	uint16_t acceptSocket(uint16_t index);
 	//关闭索引为index的socket
-	void closeSocket(int index);
+	void closeSocket(uint16_t index);
 	//接收数据
-	void recvData(int index);
+	void recvData(uint16_t index);
 	//发送数据
-	void sendData(int index, uint8_t* data, uint16_t length, uint16_t flag);
+	void sendData(uint16_t index, uint8_t* data, uint16_t length, uint16_t flag);
 	//关闭所有socket
 	void close();
 	//发生异常关闭socket
-	void errorClose(int index);
+	void errorClose(uint16_t index);
 
 	//房间管理类
 	shared_ptr<House> house;
